+++ userT/management/commands/studypdfcloseout.py
@@ -12,17 +12,22 @@ class Command(BaseCommand):
     def handle(self, *args,**options):
 
         studies = blgetAllStudies()
-        shutil.rmtree(pdfbystudy)
+        if os.path.isdir(pdfbystudy):
+            shutil.rmtree(pdfbystudy)
         for eachstudy in studies:
             StudyName = eachstudy.StudyName
-            makesubfolders = os.makedirs(pdfbystudy+f'{StudyName}',exist_ok=True) 
+            studypath = pdfbystudy+StudyName
+            studyfolder = (studypath+"/")
+            excelpath = studypath+'.csv'
+            makesubfolders = os.makedirs(studypath,exist_ok=True)   
             studydetails = ActionItems.objects.filter(StudyName__StudyName = StudyName)
             studydetailsfilter = studydetails.filter(QueSeries = 99).values()
             objactionitemsfk = blannotatefktomodel(studydetailsfilter)
             dfstudy = pd.DataFrame(objactionitemsfk)
             dfstudyfilter = dfstudy.iloc[:,:2]
-            dfstudyfilter.to_csv(pdfbystudy+f'{StudyName}'+'.csv')
-            studyfolder = (pdfbystudy+f'{StudyName}'+"/")
-            zipname = pdfbystudy+f'{StudyName}'
-            returnzipfile = blbulkdownload(objactionitemsfk,studyfolder,zipname)
+            dfstudyfilter.to_csv(excelpath)
+            returnzipfile = blbulkdownload(objactionitemsfk,studyfolder,studypath)
+        change_permissions_recursive(pdfbystudy, 0o770)     # 770 is linux permission
+        change_group_recursive(pdfbystudy, 1000, 1004)      # 1000 is bitnami in linux, 1004 is varwwwusers (user group)
+
         return returnzipfile
